1.Ultimo commit na Summary (entendi errado e a lógica é para gerar gráficos de improdutividade nas barras (e isso muda consideravelmente a lógica, logo vou reescrever do 0))
2.Mudanca pontual no Context (deleta somente se houver o arquivo de db)

3.Usando o Init parcial na main (deleta o estado atual, insere tudo de novo)

+++ src/sapo/Context.hx
@@ -53,7 +53,8 @@ class Context {
 			Manager.cnx.close();
 			Manager.cnx = null;
 		}
-		sys.FileSystem.deleteFile(DBPATH);
+		if(sys.Filesystem.exists(DBPATH))
+			sys.FileSystem.deleteFile(DBPATH);
 		init();
 
 		Manager.cnx.request("BEGIN");
